Bug 1507923 - Extract some useful git-related bits of code into a new module
