+++ tools/src/git_ops.rs
@@ -1,0 +1,45 @@
+use std::path::Path;
+
+use git2::{Commit, Repository, TreeEntry};
+
+use config::GitData;
+
+// Helpers to do things with git2
+
+fn latin1_to_string(bytes: Vec<u8>) -> String {
+    bytes.iter().map(|&c| c as char).collect()
+}
+
+pub fn decode_bytes(bytes: Vec<u8>) -> String {
+    match String::from_utf8(bytes.clone()) {
+        Ok(s) => s,
+        Err(_) => {
+            latin1_to_string(bytes)
+        }
+    }
+}
+
+pub fn read_blob_entry(repo: &Repository, entry: &TreeEntry) -> String {
+    let blob_obj = entry.to_object(repo).unwrap();
+    let blob = blob_obj.as_blob().unwrap();
+    let mut content = Vec::new();
+    content.extend(blob.content());
+    decode_bytes(content)
+}
+
+pub fn get_blame_lines(git_data: Option<&GitData>, blame_commit: &Option<Commit>, path: &str) -> Option<Vec<String>> {
+    match (git_data, blame_commit) {
+        (Some(&GitData {blame_repo: Some(ref blame_repo), ..}), &Some(ref blame_commit)) => {
+            let blame_tree = blame_commit.tree().ok()?;
+
+            match blame_tree.get_path(Path::new(path)) {
+                Ok(blame_entry) => {
+                    let blame_data = read_blob_entry(blame_repo, &blame_entry);
+                    Some(blame_data.lines().map(str::to_string).collect::<Vec<_>>())
+                },
+                Err(_) => None,
+            }
+        },
+        _ => None,
+    }
+}
