+++ tools/src/lib.rs
@@ -17,6 +17,7 @@ pub mod languages;
 pub mod format;
 pub mod tokenize;
 pub mod links;
+pub mod git_ops;
 
 pub fn find_source_file(path: &str, files_root: &str, objdir: &str) -> String {
     if path.starts_with("__GENERATED__") {
